re-refactor structure of feature_server
